// src/shared/types.ts
export type Widget = {
  id?: string;
  /** Does it have a background? */
  hasBackground?: boolean;
};
// ...
export type Icon = TablerIcon;

export type CallToActionType = {
  text?: string;
  href: string;
  icon?: Icon;
  targetBlank?: boolean;
};

export type LinkOrButton = {
  callToAction?: CallToActionType;
  containerClass?: string;
  linkClass?: string;
  iconClass?: string;
};

export type Button = {
  title: string;
  type: 'button' | 'submit' | 'reset';
};

export type Input = {
  type: string;
  label?: string;
  value?: string;
  name?: string;
  autocomplete?: string;
  placeholder?: string;
};

export type Textarea = {
  cols?: number;
  rows?: number;
  label?: string;
  name: string;
  placeholder?: string;
};

export type Checkbox = {
  label: string;
  value: string;
};

export type Radio = {
  label: string;
};

export type RadioBtn = {
  label?: string;
  radios: Array<Radio>;
};

export type SmallForm = {
  icon?: Icon;
  input: Input;
  btn: Button;
};

export type FormProps = {
  title?: string;
  description?: string;
  inputs: Array<Input>;
  radioBtns?: RadioBtn;
  textarea?: Textarea;
  checkboxes?: Array<Checkbox>;
  btn: Button;
  btnPosition?: 'center' | 'right' | 'left';
  containerClass?: string;
};

export type Image = {
  link?: string;
  src: string | StaticImageData;
  alt: string;
};

export type Item = {
  title?: string | boolean | number;
  description?: string | Array<string>;
  href?: string;
  form?: SmallForm;
  icon?: Icon;
  callToAction?: CallToActionType;
};
// ...
export type CallToActionProps = Widget & {
  title: string;
  subtitle: string;
  callToAction?: CallToActionType;
  items?: Array<Item>;
};